World ID
Defaults to 12, if the config file is already set and another mod is
already using that world id it will reset to -1. (It should throw an
exception in this case.)

If the config file is not or the value is -1  and world type 12 is not
free it will attempt to find a free world id. If it can't it presently
puts a message in the log. (It should throw an exception.)